Fix unblocking or non-pending takes.
If you poll! or try taking within alt!/alts! on a stream, we
need to stash the deferred away so that if the first op was
a poll, or wasn't completed in an alt op, subsequent takes will
get that value, instead of attempting to take! from the stream
again.

I think this probably won't work if you are doing multiple
blocking take operations at once -- in that case each take
would get the realized value at the same time. I'll think more
on the implementation of this for streams.

* src/intake/core/async.clj (pending-reads): new var.
  (IEventSource.take!): for blockable handlers, see if we stashed
  a previous take, and stash any new takes we make. For unblocking
  takes, use try-take! with a zero timeout.
  (IDeferred.take!): handle default values and timeouts.
* test/intake/core/async_test.clj: update tests. Add overall timeout.
* project.clj: update core.async version.